change the image name while testing

+++ hashencoder/src/hashencoder.cu
@@ -22,7 +22,7 @@
 // requires CUDA >= 10 and ARCH >= 70
 // this is very slow compared to float or __half2, do not use!
 static inline  __device__ at::Half atomicAdd(at::Half *address, at::Half val) {
-  return atomicAdd(reinterpret_cast<__half*>(address), val);
+  return atomicAdd(reinterpret_cast<float*>(address), val);
 }
